+++ packages/zowe-explorer/src/webviews/src/table-view/App.tsx
@@ -9,8 +9,8 @@
  *
  */
 
-export * from "./utils/TableBuilder";
-export * from "./utils/TableMediator";
-export * from "./TableView";
-export * from "./TableViewProvider";
-export * from "./WebView";
+import { TableView } from "./TableView";
+
+export function App() {
+  return <TableView />;
+}